refactored code to reuse JSONPredicate class from standalone but moved it to common
